apps/bot/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.0
redis==5.2.1
transitions==0.9.2
httpx==0.28.1
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.25.0
pyyaml==6.0.2
python-telegram-bot==21.10
fakeredis==2.26.2
